Move all vector tests to NativeVectorTestsBase.cs
